Cleanup and rename LegacyParsingInterceptor to JSONResponseParsingInterceptor
