+++ Tests/ApolloTests/InterceptorTests.swift
@@ -88,7 +88,7 @@ class InterceptorTests: XCTestCase {
           MaxRetryInterceptor(maxRetriesAllowed: self.retryCount),
           self.testInterceptor,
           NetworkFetchInterceptor(client: self.mockClient),
-          LegacyParsingInterceptor(),
+          JSONResponseParsingInterceptor(),
         ]
       }
     }
@@ -134,7 +134,7 @@ class InterceptorTests: XCTestCase {
       func interceptors<Operation: GraphQLOperation>(for operation: Operation) -> [ApolloInterceptor] {
         [
           NetworkFetchInterceptor(client: self.mockClient),
-          LegacyParsingInterceptor(),
+          JSONResponseParsingInterceptor(),
         ]
       }
     }
@@ -154,7 +154,7 @@ class InterceptorTests: XCTestCase {
         XCTFail("This should not have succeeded")
       case .failure(let error):
         switch error {
-        case LegacyParsingInterceptor.LegacyParsingError.couldNotParseToLegacyJSON(let data):
+        case JSONResponseParsingInterceptor.JSONResponseParsingError.couldNotParseToLegacyJSON(let data):
           XCTAssertTrue(data.isEmpty)
         default:
           XCTFail("Unexpected error type: \(error.localizedDescription)")
@@ -183,7 +183,7 @@ class InterceptorTests: XCTestCase {
         [
           NetworkFetchInterceptor(client: self.mockClient),
           ResponseCodeInterceptor(),
-          LegacyParsingInterceptor()
+          JSONResponseParsingInterceptor()
         ]
       }
     }
@@ -203,7 +203,7 @@ class InterceptorTests: XCTestCase {
         XCTFail("This should not have succeeded")
       case .failure(let error):
         switch error {
-        case LegacyParsingInterceptor.LegacyParsingError.couldNotParseToLegacyJSON(let data):
+        case JSONResponseParsingInterceptor.JSONResponseParsingError.couldNotParseToLegacyJSON(let data):
           XCTAssertTrue(data.isEmpty)
         default:
           XCTFail("Unexpected error type: \(error.localizedDescription)")
@@ -239,7 +239,7 @@ class InterceptorTests: XCTestCase {
         [
           NetworkFetchInterceptor(client: self.mockClient),
           ResponseCodeInterceptor(),
-          LegacyParsingInterceptor(),
+          JSONResponseParsingInterceptor(),
         ]
       }
     }
